% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/arrival.process.R
\name{arrival.process}
\alias{arrival.process}
\title{Generates patient arrival data}
\usage{
arrival.process(nr.datasets = 1, seed = NULL, rates.by.disease,
  mutants.by.disease = NULL, mutants.by.module = NULL, by.module = FALSE)
}
\arguments{
\item{nr.datasets}{Number of arrival-datasets to be generated.}

\item{seed}{Set seed equela to \code{seed}.}

\item{rates.by.disease}{A vector or matrix of arrival rates for K disease (columns) and M subpopulations (rows),
for a trial M Biomarker subpopulations.}

\item{mutants.by.disease}{A vector or matrix of total patient accrual  by disease (columns) and subpopulation (rows).}

\item{mutants.by.module}{A scalar or vector of total patient accrual  by subpopulation.}

\item{by.module}{If true simulate patient total patient accrual by model (fixed) and
then simulated for the \code{mutants.by.module[m]} patients on subpopulation the disease type.}
}
\value{
Returns a list of nr.datasets  matrices, where each row indicates a patient.
The 1th, 2th and 3th column contains the arrival-time, disease and subpopulation for each patient.
}
\description{
Takes the a vector of arrival rates and the number of patients to be accruled by disease and returns patients arrival data by time (in weeks).
}
\examples{
## one module with 3 cancer
rates.by.disease = c(1,2,3)
mutants.by.disease = c(10, 20, 10)
arrival.process(nr.datasets=10, seed=1121, rates.by.disease, mutants.by.disease)

## two modules with 3 cancer
rates.by.disease = cbind(c(1,0,3), c(0,3,3))
mutants.by.disease = cbind(c(10, 0, 10), c(0, 10, 10))
arrival.process(nr.datasets=10, seed=1121, rates.by.disease, mutants.by.disease)
}
\author{
Steffen Ventz  \email{ventzer@yahoo.de}
}

