% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BvM.approximation.R
\name{BvM.approximation}
\alias{BvM.approximation}
\title{Posterior for the Bayesian Probit Model}
\usage{
BvM.approximation(theta.mu = rep(0, length(Events)), m1H.prior, Events, Risk,
  maxir = 50, eps = 10^-5)
}
\arguments{
\item{theta.mu}{starting value of the algorithm}

\item{m1H.prior}{the negative Hessian of the log-prior density, which is the negative of the inverse prior covariance matrix}

\item{Events}{A vector of length \code{nrow(m1H.prior)} of sucesses}

\item{Risk}{A vector of length \code{nrow(m1H.prior)} of sucesses plus failure}

\item{maxir}{Maximum number of iterations of the algorithm}

\item{eps}{Stoping criteria, stop if the L1 distance between the parameter vector for two succesive iterations is below \code{eps}.}
}
\value{
The function returns a lsit of two object \code{theta.mode} and \code{Asymp.Var}, which are the posterior mode and the approximate Covatiance matrix of the estimated posterior mode.
}
\description{
The function compute the posterior mode of a Bayesian probit model for Gaussion prior using a Neton-Rapson Algorithm
}
\author{
Steffen Ventz  \email{ventzer@yahoo.de}
}
\keyword{internal}

