% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Cov.parameter_FCT.R
\name{Cov.parameter_FCT}
\alias{Cov.parameter_FCT}
\title{Covariance Matrix for a the Gaussian Prior}
\usage{
Cov.parameter_FCT(eligibility.array, Var.vec = c(0, 0.8, 0, 0.1, 0.05))
}
\arguments{
\item{eligibility.array}{Eligibility array, a two or three dimensional array of zeros and ones.
The entry \code{eligibility.array[d,a,m]} is one if patients with disease \code{d} in subpopulation \code{m} are eligible for
treatment with agent \code{a}.}

\item{Var.vec}{Vector of five length five of variances of the parameter
 \eqn{\alpha_{d}}, \eqn{\alpha_{d,m}}, \eqn{\beta_{a}},  \eqn{\beta_{a,m}} and \eqn{\beta_{d,a,m}}.}
}
\value{
Returns the covariance matrix of a Gausian prior for one module with variance components \code{var.alpha}, \code{var.beta} and \code{var.gamma}.
}
\description{
Covariance Matrix for the Gaussian Prior for a Bayesian Hierarchical Probit Regression Model
}
\examples{
eligibility.array      = array( dim=c(4,4,2))
eligibility.array[,,1] = cbind( c(0,1,0,1), rep(0,4), c(0,1,0,1), c(0,0,0,1))
eligibility.array[,,2] = cbind( c(1,1,1,1), c(1,0,1,1), c(0,1,1,0), c(1,0,0,1))
dimnames(eligibility.array) = list( paste0("disease_", 1:4),
c("control", paste0("agent", 1:3)),
paste0("marker", 1:2))

## Exa_1: one module only
Cov.parameter_FCT(eligibility.array[,,2], c(0, .8, 0, .1, .05))

## Exa_2: two subpopulations
Cov.parameter_FCT(eligibility.array, c(.1, .8, .1, .1, .05))
}
\author{
Steffen Ventz  \email{ventzer@yahoo.de}
}

