% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get.eligibility.vec.id.R
\name{get.eligibility.vec.in}
\alias{get.eligibility.vec.in}
\title{Internal eligibility objects}
\usage{
get.eligibility.vec.in(eligibility.array)
}
\arguments{
\item{eligibility.array}{Eligibility array, an array of zeros and one where
\code{eligibility.array[d,a,m]=1} indicated that agents \code{a} is a treatment option for  disease/subpopulation
\code{d,m}.}
}
\value{
List of objects used to simulate the trial.
}
\description{
Takes \code{eligibility.array}, the eligibility array and
computes a list of objects required internally to simulate the trial.
}
\examples{
### 3 cancer, 5 agents and 2 modules
eligibility.array = array(1, c(3, 5, 2))
eligibility.array[3,3:4,2] = 0
eligibility.array[1,2:3,1] = 0
get.eligibility.vec.in(eligibility.array)
}
\author{
Steffen Ventz \email{ventzer@yahoo.de}
}
\keyword{internal}

