% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.outcome.R
\name{get.outcome}
\alias{get.outcome}
\title{Generate response to treatment}
\usage{
get.outcome(seed = 1111, rate, Arrival = NULL)
}
\arguments{
\item{seed}{Fix seed at  \code{seed}.}

\item{rate}{An array of response probabilities, 
where the component (d,a,m) denotes probability of a positive respons to treatment a for a patient with disease d and  subpopulation m.}

\item{Arrival}{A list of arrival data generated with arrival_process.}
}
\value{
A list of outcome matrices, where the component (n,a) in each matrix denotes the response-outcome under treatment a for patient n.
}
\description{
The function simulates the response-to-treatment outcome for  \code{N} patients for
under treatment regimen conditional on the disease type and subpopulation.
}
\examples{

mutants.by.disease = rbind(c(30, 30), c(30, 30))
rates.by.disease   = rbind(c(1.2, 0.7), c(1.0, 0.5))
Arrival  = arrival.process(nr.datasets=1, seed=1121, rates.by.disease, mutants.by.disease)

# two cancer, 5 agents and 2 subpopulations
rate      = array(0, c(2, 4, 2))
rate[,,1] = matrix(c(.3, .5, .3, .3), 2, 4, byrow=TRUE)
rate[,,2] = matrix(c(.2, .4, .3, .2), 2, 4, byrow=TRUE)


get.outcome(121, rate=rate, Arrival=Arrival)


}
\author{
Steffen Ventz  \email{ventzer@yahoo.de}
}
