% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.Store.Trial.Info.R
\name{get.Trial.Results}
\alias{get.Trial.Results}
\title{Create storage for trial data}
\usage{
get.Trial.Results(Dim, Check, N)
}
\arguments{
\item{Dim}{A vector of length 3, specifying the number of number of diease, agents and modules.}

\item{Check}{a Boolen vector of length 5 to 7.
Only the first 6 components are required and have to be specify (with \code{TRUE} or  \code{FALSE}).
The second, third and fourth component specify whether patient accrual,
the empirical response probability and the efficacy statistics should be traced.
The fifth components specifies in addition if the the efficacy statistics should be specified for all agents by sidese and module
or for one agent and a specific disease/module only.
The sixed component specifies whether normal random variables should be samples for the randomization procedure.}

\item{N}{The (expected) number of patients to be observed during the trial.}
}
\value{
If \code{Check[2]=TRUE} An allocation-array of dimension
}
\description{
The function create a set of arrays and matrices which are used within the process
of trials simulation and and inference.
}
\author{
Steffen Ventz  \email{ventzer@yahoo.de}
}
\keyword{internal}
