% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.vec.from.mat.R
\name{get.vec.from.mat}
\alias{get.vec.from.mat}
\title{Transform response and event matrix into two columns of one matrix}
\usage{
get.vec.from.mat(Resp_Events, Resp_RiskPop, eligibility.vec.id)
}
\arguments{
\item{Resp_Events}{array of observed response events (sucesses) by disease (rows) and agents (columns)}

\item{Resp_RiskPop}{array of observed events (sucesses and failure) by disease (rows) and agents (columns)}

\item{eligibility.vec.id}{Boolen vector generated with (see general vignette)}
}
\value{
A matrix of observed data. First column equals the vector observed respons events. Second column equals the vector of observed number of events.
}
\description{
The function takes the response and event array and stacks all columns below each other to compute one big vector of response counts and one vector of event counts.
}
\author{
Steffen Ventz  \email{ventzer@yahoo.de}
}
\keyword{internal}
