% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gradient.Hessian.ll.R
\name{gradient.Hessian.ll}
\alias{gradient.Hessian.ll}
\title{Hessian and Gradient for a Probit model}
\usage{
gradient.Hessian.ll(theta, Events, Risk)
}
\arguments{
\item{theta}{parameter vector of a logit model}

\item{Events}{vector of observed response events (sucesses). First row of the matrix created with \code{get.vec.from.mat} (see vignette for details)}

\item{Risk}{vector of observed events (sucesses and failure). Second row of the matrix created with \code{get.vec.from.mat} (see vignette for details)}
}
\value{
A list of two elements (H, G) where H anf G are the Hessian and Gradient of the log-likelihood under a Probit model
}
\description{
The function computes the Hessian and Gradient of the log-likelihood under a Probit model
}
\author{
Steffen Ventz  \email{ventzer@yahoo.de}
}
\keyword{internal}

