% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/id.fct.R
\name{id.fct}
\alias{id.fct}
\title{Three dimensional row-major order}
\usage{
id.fct(a, m, d, Dim)
}
\arguments{
\item{a}{control \eqn{a=1} and experimental agent  \eqn{a>1} index}

\item{m}{Biomarker index}

\item{d}{disease indes}

\item{Dim}{Three dimensional vector with components total number of disease, agents and modules.}
}
\description{
Compute the row-major order for a three dimensional  array \eqn{(a,m,d)}
}
\author{
Steffen Ventz  \email{ventzer@yahoo.de}
}
\keyword{internal}

