% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initial.rand_summary.R
\name{initial.rand.summary}
\alias{initial.rand.summary}
\title{Generates an array of current trial information}
\usage{
initial.rand.summary(eligibility.array)
}
\arguments{
\item{eligibility.array}{An array with elements \code{TRUE, FALSE} which indicates the eligibility of an agent for each disease and subpopulation.}
}
\value{
The function returns an array, say rand_summary, of dimention (m,n,5).
\code{rand_summary[,,i,m]} gives
the status of each agenty by disease in module m.
Encoding: 1=active, 0=dropped for futility, 2=graduated for efficacy, -1=not eligible.
\code{rand_summary[,,2,m]} gives
for agents that were dropped/graduated the stopping time by disease (by columns) in module m. .
\code{rand_summary[,,3,m]} gives
the current randomization probabilities by disease (by columns).
\code{rand_summary[,,4,m]} and
rand_summary[,,5,m] return the futility and efficacy statistics by agent (row)
and disease (column).
}
\description{
The function computes an array.
}
\examples{
## 3 cancer and 5 agents eligible for each cancer
eligibility.TRUE = matrix(TRUE, 3, 5)
initial.rand.summary(eligibility.TRUE)

## 3 cancer and 5 agents, 2 modules
eligibility.TRUE = array(TRUE, c(3, 5, 2))
eligibility.TRUE[3,3:4,2] = TRUE
eligibility.TRUE[1,2:3,1] = TRUE
initial.rand.summary(eligibility.TRUE)

#@export
}
\author{
Steffen Ventz  \email{ventzer@yahoo.de}
}
