% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Plot.Prior.R
\name{Plot.prior}
\alias{Plot.prior}
\title{Plot the prior density for the response rate}
\usage{
Plot.prior(filename.with.path = "Plot", Var, plot.figure = NULL)
}
\arguments{
\item{filename.with.path}{A character string of path and filename.
The function saves all three figures by default as
\code{filename.with.path, ".Joined.pdf"},
\code{filename.with.path, ".Conditional.dd.pdf"} and
\code{filename.with.path, ".Conditional.mm.pdf"}.}

\item{Var}{Vector of prior variance parameter \eqn{(W_1, W_2, W_3, W_4, W_5)=Var(\alpha_{d}, \alpha_{d, m}, \beta_{a}, \beta_{a,m}, \beta_{a,m,d})}.}

\item{plot.figure}{If \code{plot.figure=1,2,3} the function plots
(1) the joined probability of \eqn{p_{d,a,m}, p_{d',a,m'}}
(2) conditional prior density of \eqn{p_{d,a,m}, p_{d',a,m}}  given  \eqn{p_{d,0,m}, p_{d',0,m}} or
(3) conditional prior density of \eqn{p_{d,a,m}, p_{d,a,m'}}  given  \eqn{p_{d,0,m}, p_{d',0,m'}}.}
}
\description{
Plot the prior density for the response rate
}
\details{
The function prepares plots the prior density of \eqn{p_{d,a,m}} for \eqn{a>0},
joined density of \eqn{p_{d,a,m}, p_{d',a,m'}}
and the conditional prior density of   \eqn{p_{d,a,m}, p_{d',a,m'}} given  \eqn{p_{d,0,m}, p_{d',0,m'}}.
All three figures are saved as \code{filename.with.path, ".Joined.pdf"},
\code{filename.with.path, ".Conditional.dd.pdf"} and \code{filename.with.path, ".Conditional.mm.pdf"}.
}
\examples{
Plot.prior(filename.with.path="Plot",
Var= c(V1=.1, V2=.1, W1=.3, W2=.1, W3=.05))
}

