% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Response_delay_data_module_FCT.R
\name{Response_delay_data_FCT}
\alias{Response_delay_data_FCT}
\title{Generates a matrix to store data not jet available}
\usage{
Response_delay_data_FCT()
}
\value{
Returns a matrix of dimension (5,2). Each column denotes a patient.
Rows 1 to 5 indicate the disease, agent, time at which outcome will be available, the respone to treatment and the module.
#
}
\description{
Generates a matrix to store data not jet available during the trial.
}
\examples{
## Response_delay_data_module_FCT()
}
\author{
Steffen Ventz  \email{ventzer@yahoo.de}
}
\keyword{internal}

