% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Response_Events_Risk_module_FCT.R
\name{Response_Events_Risk_FCT}
\alias{Response_Events_Risk_FCT}
\title{Generates a matrix to store patient data}
\usage{
Response_Events_Risk_FCT(Dim)
}
\arguments{
\item{Dim}{A vector of 3 positive integer, representing the number of disease, agents and modules.}
}
\value{
Returns an array of dimension (D,A,M), where
D is the total number of disease in the trial,
A is the total number of agents in the trial and 
M is the total number of modules in the trial.
}
\description{
The functions uses the object \code{Module_Info} created with the function 
\code{Module_Info_FCT} to create an array for storing the response data.
}
\author{
Steffen Ventz  \email{ventzer@yahoo.de}
}
\keyword{internal}
