% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_Data.module.R
\name{update_Data_Fct}
\alias{update_Data_Fct}
\title{Update availabe data}
\usage{
update_Data_Fct(T.i = 0, Resp_Delay = NULL, Resp_RiskPop = NULL,
  Resp_Events = NULL, Dim = 1, Initial = FALSE)
}
\arguments{
\item{T.i}{Current time of the trial}

\item{Resp_Delay}{The current number of not jet visible data.}

\item{Resp_RiskPop}{The current number of observed events by agents and disease}

\item{Resp_Events}{The current number of responders observed by agents and disease}

\item{Dim}{A vector of 3 positive integer, representing the number of disease, agents and modules.}

\item{Initial}{Boolen vector indicatin whether to initiate \code{Resp_Events}, \code{Resp_RiskPop} and \code{Resp_Delay}.}
}
\value{
Returns a list of elements \code{Resp_Delay}, \code{Resp_RiskPop}, \code{Resp_Events}, \code{Any.change}.
If new data become available, then \code{Any.change=TRUE} and
\code{Resp_Delay}, \code{Resp_RiskPop}, \code{Resp_Events} are updated according newly available data.
}
\description{
## The function checks if events not visible to the trial before become available at time T.i.
The number of observed events and responders by disease are update by the newly available data.
}
\examples{
update_Data_Fct(T.i=0, NULL, NULL, NULL, Dim=c(2,4,2), Initial=TRUE)
}
\author{
Steffen Ventz  \email{ventzer@yahoo.de}
}
\keyword{internal}
