% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/update_Posterior.R
\name{updatePosterior}
\alias{updatePosterior}
\title{Compute posterior for a hierarchical trial design}
\usage{
updatePosterior(Resp_Events, Resp_RiskPop, Prior, Design = c(2, 3),
  alpha.historical = NULL, p.historical = NULL)
}
\arguments{
\item{Resp_Events}{Array of observed response/succeses for the control and experimental agents for all disease by subpopulations.}

\item{Resp_RiskPop}{Array of observed outcome for the control and experimental agents for all disease by subpopulations.}

\item{Prior}{A list of prior parameter generated with \code{Prior.list.Fct}}

\item{Design}{A vector of length 2 with specifies the experimental design of the trial.
For \code{Design[1] = 1} or \code{Design[1] = 2} the function returns
functions of the posterior for a Phase II Biomarker-finding or Biomarker-stratified design
For \code{Design[2] = 1, 2, 3} specifise the presence of an (1) active, (2) historical control arm, and (3) the absence of a control arm.}

\item{alpha.historical}{A matrix of the inverse normal of the historical/null response probabilities by disease (rows) and subpopulations (columns), i.e.
\code{pnorm(alpha.historical[d,m])} should be equal to \code{p.historical[d,m]}.}

\item{p.historical}{A matrix of  historical/null response probabilities by disease (rows) and subpopulations (columns).
 Note that it is required that \code{pnorm(alpha.historical[d,m])} should be equal to \code{p.historical[d,m]}.}
}
\value{
A list of three elements;
(i) \code{Rand.Pr} array of non-amplified randomization statistics,
(ii) \code{Fut.Pr} array of  futility statistics,
(iii) \code{Eff.stat} array of  efficacy statistics.
}
\description{
Compoute functions of the posterior require for the adaptive randomization and for applying early the decision rules.
}
\details{
For \code{Design[1] = 1} or \code{Design[1] = 2} the function returns
functions of the posterior for a Phase II Biomarker-finding or  Biomarker-stratified design
For \code{Design[2] = 1, 2, 3} specifise the presence of an (1) active, (2) historical control arm, and (3) the absence of a control arm.
}
\examples{
### example1 : 2 cancer, 5 experimental agents and 2 modules, active control
 eligibility.array        = array(1, c(2, 5, 2))
 Resp_Events              = Resp_RiskPop = array(0, c(2, 5, 2))
 eligibility.array[2,2,2] = 0
 eligibility.array[1,,1]  = 0
 Dim                      = dim(eligibility.array)

 ### response rates
 rate                      = array(NA, Dim)
 rate[,,1]                 = matrix(c(.3, .3, .5, .3, .3, .3, .3, .2, .4, .3), 2, 5, byrow=TRUE)
 rate[,,2]                 = matrix(c(.3, .2, .4, .3, .3, .3, .3, .2, .4, .3), 2, 5, byrow=TRUE)

 ### generate data
 Resp_Events = Resp_RiskPop = array(0, dim=Dim)

 for(d in 1:Dim[1]) for(a in 1:Dim[2])
 Resp_RiskPop[d,a,] = rpois(Dim[3], 150)
 for(d in 1:Dim[1]) for(a in 1:Dim[2])
 Resp_Events[d,a,] = rbinom(Dim[3], Resp_RiskPop[d,a,], rate[d,a,])

 Resp_Events [eligibility.array==0] = 0
 Resp_RiskPop[eligibility.array==0] = 0

 Prior = Prior.list.Fct(eligibility.array=eligibility.array,
              Var.vec = c(alpha1=1, alpha2=2, beta1=.2, beta2=.1, beta3=.05))
 updatePosterior(Resp_Events, Resp_RiskPop, Prior, Design = c(2,1)) ## SSD
 updatePosterior(Resp_Events, Resp_RiskPop, Prior, Design = c(1,1)) ## SFD


### example2 : 2 cancer, 5 experimental agents and 2 modules, no/historical control
 eligibility.array[,1,]             = 0

 Prior = Prior.list.Fct(eligibility.array=eligibility.array,
                 Var.vec = c(alpha1=0, alpha2=0, beta1=.2, beta2=.1, beta3=.05))
 Resp_Events[eligibility.array==0]  = Resp_RiskPop[eligibility.array==0] =0


 updatePosterior(Resp_Events, Resp_RiskPop, Prior, Design = c(1,3),
                  alpha.historical   = matrix(qnorm(.3), nrow=Dim[1], ncol=Dim[3]),
                  p.historical       = matrix(.3, nrow=Dim[1], ncol=Dim[3]) ) ## SFD no control

 updatePosterior(Resp_Events, Resp_RiskPop,Prior, Design = c(2,2),
                  alpha.historical   = matrix(qnorm(.3), nrow=Dim[1], ncol=Dim[3]),
                  p.historical       = matrix(.3, nrow=Dim[1], ncol=Dim[3]) )
}
\author{
Steffen Ventz  \email{ventzer@yahoo.de}
}

