% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Z.statistics.mat.R
\name{Z.statistics.arr}
\alias{Z.statistics.arr}
\title{Array-valued Z-statistics}
\usage{
Z.statistics.arr(N.0 = NULL, N.a, R.0 = NULL, R.a, One.sample = FALSE,
  p.historical = NULL)
}
\arguments{
\item{N.0}{Matrix of observed outcome for the standard of care for all disease by subpopulations.}

\item{N.a}{Array of observed outcome for the experimental agents for all disease by subpopulations.}

\item{R.0}{Matrix of observed response/succeses for the standard of care for all disease by subpopulations.}

\item{R.a}{Array of observed response/succeses for the experimental agents for all disease by subpopulations.}

\item{One.sample}{If \code{TRUE}, use z-statistics for  the onse-sided  null hypothesis \eqn{ p_{d,a,m} \le } \code{p.historical[d,m]}.}

\item{p.historical}{An matrix  of probabilityies corresponding to the null respons probability \eqn{p_{d,a,m} \le } \code{p.historical[d,m]}.}
}
\description{
Computes the coordinate-wise Z-statistics
}
\details{
The function takes the array of observed events and number of observed outcomes,
and computes an array of Z-statistics for Binomial data.
}
\author{
Steffen Ventz  \email{ventzer@yahoo.de}
}
\keyword{internal}

