% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/initial.rand_pr.R
\name{initial.rand_pr}
\alias{initial.rand_pr}
\title{Generates initiual balanced randomization probabilities}
\usage{
initial.rand_pr(eligibility.array)
}
\arguments{
\item{eligibility.array}{An array with elements \code{TRUE, FALSE} which indicates the eligibility of an agent for each disease and subpopulation.}
}
\value{
An array of randomization probabilities.
The element (d,a,m) denotes the initial probability if randomizing a patient with disease d in subgroup m to agents a.
This number is zero if the \code{eligibility.array[d,a,m]=0}.
#
}
\description{
The function computes uniform randomization rates among all eligible agents.
}
\examples{
#
### 3 cancer, 5 agents and 1 modules
#initial.rand_pr(matrix(TRUE, 3, 4))
#
#### 3 cancer, 5 agents and 2 modules
#eligibility.array = array(TRUE, c(3, 5, 2))
#eligibility.array[3,3:4,2] = FALSE
#eligibility.array[1,2:3,1] = FALSE
#initial.rand_pr(eligibility.array)
#
}
\author{
Steffen Ventz  \email{ventzer@yahoo.de}

Steffen Ventz
}
\keyword{internal}

